Move get_run_status.py -> get_status.py
This replaces the old get_status.py script
